store_api::data_source
Trait DataSource

pub trait DataSource {
    // Required method
    fn get_stream(
        &self,
        request: ScanRequest,
    ) -> Result<SendableRecordBatchStream, BoxedError>;
}

This trait represents a common data source abstraction which provides an interface for retrieving data in the form of a stream of record batches.

Required Methods

fn get_stream( &self, request: ScanRequest, ) -> Result<SendableRecordBatchStream, BoxedError>

Retrieves a stream of record batches based on the provided scan request.