Trait store_api::logstore::entry::Entry

pub trait Entry: Send + Sync {
    type Error: ErrorExt + Send + Sync;
    type Namespace: Namespace;

    // Required methods
    fn data(&self) -> &[u8] ;
    fn id(&self) -> Id;
    fn namespace(&self) -> Self::Namespace;
}

Entry is the minimal data storage unit through which users interact with the log store. The log store implementation may have larger or smaller data storage unit than an entry.

Required Associated Types

type Error: ErrorExt + Send + Sync

type Namespace: Namespace

Required Methods

fn data(&self) -> &[u8]

Returns the contained data of the entry.

fn id(&self) -> Id

Returns the id of the entry. Usually the namespace id is identical with the region id.

fn namespace(&self) -> Self::Namespace

Returns the namespace of the entry.