Trait store_api::manifest::MetaAction

pub trait MetaAction: Serialize + DeserializeOwned + Send + Sync + Clone + Debug {
    type Error: ErrorExt + Send + Sync;

    // Required methods
    fn set_protocol(&mut self, action: ProtocolAction);
    fn set_prev_version(&mut self, version: ManifestVersion);
    fn encode(&self) -> Result<Vec<u8>, Self::Error>;
    fn decode(
        bs: &[u8],
        reader_version: ProtocolVersion
    ) -> Result<(Self, Option<ProtocolAction>), Self::Error>;
}

The action to alter metadata

Required Associated Types

type Error: ErrorExt + Send + Sync

Required Methods

fn set_protocol(&mut self, action: ProtocolAction)

Set a protocol action into meta action

fn set_prev_version(&mut self, version: ManifestVersion)

Set previous valid manifest version.

fn encode(&self) -> Result<Vec<u8>, Self::Error>

Encode this action into a byte vector

fn decode( bs: &[u8], reader_version: ProtocolVersion ) -> Result<(Self, Option<ProtocolAction>), Self::Error>

Decode self from byte slice with reader protocol version, return error when reader version is not supported.

Object Safety

This trait is not object safe.