store_api
Module mito_engine_options

Option keys for the mito engine. We define them in this mod so the create parser can use it to validate table options.

Constants

APPEND_MODE_KEY

Option key for append mode.

COMPACTION_TYPE

Option key for compaction type.

COMPACTION_TYPE_TWCS

TWCS compaction strategy.

MERGE_MODE_KEY

Option key for merge mode.

REMOTE_COMPACTION

Option key for twcs remote compaction.

TTL_KEY

Option key for TTL(time-to-live)

TWCS_FALLBACK_TO_LOCAL

Option key for twcs fallback to local.

TWCS_MAX_ACTIVE_WINDOW_FILES

Option key for twcs max active window files.

TWCS_MAX_ACTIVE_WINDOW_RUNS

Option key for twcs max active window runs.

TWCS_MAX_INACTIVE_WINDOW_FILES

Option key for twcs max inactive window files.

TWCS_MAX_INACTIVE_WINDOW_RUNS

Option key for twcs max inactive window runs.

TWCS_MAX_OUTPUT_FILE_SIZE

Option key for twcs max output file size.

TWCS_TIME_WINDOW

Option key for twcs time window.

Functions

is_mito_engine_option_key

Returns true if the key is a valid option key for the mito engine.