Module store_api::region_engine

Region Engine’s definition

Structs

GrantedRegion

PartitionRange

Represents one data range within a partition

ScannerProperties

Properties of the RegionScanner.

SinglePartitionScanner

A RegionScanner that only scans a single partition.

Enums

RegionRole

ScannerPartitioning

Output partition properties of the RegionScanner.

SetReadonlyResponse

The result of setting readonly for the region.

Traits

RegionEngine

RegionScanner

A scanner that provides a way to scan the region concurrently. The scanner splits the region into partitions so that each partition can be scanned concurrently. You can use this trait to implement an ExecutionPlan.

Type Aliases

BatchResponses

RegionEngineRef

RegionScannerRef