Trait store_api::region_engine::RegionEngine

pub trait RegionEngine: Send + Sync {
    // Required methods
    fn name(&self) -> &str;
    fn handle_request<'life0, 'async_trait>(
        &'life0 self,
        region_id: RegionId,
        request: RegionRequest
    ) -> Pin<Box<dyn Future<Output = Result<RegionHandleResult, BoxedError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_query<'life0, 'async_trait>(
        &'life0 self,
        region_id: RegionId,
        request: ScanRequest
    ) -> Pin<Box<dyn Future<Output = Result<SendableRecordBatchStream, BoxedError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_metadata<'life0, 'async_trait>(
        &'life0 self,
        region_id: RegionId
    ) -> Pin<Box<dyn Future<Output = Result<RegionMetadataRef, BoxedError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn region_disk_usage<'life0, 'async_trait>(
        &'life0 self,
        region_id: RegionId
    ) -> Pin<Box<dyn Future<Output = Option<i64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn stop<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), BoxedError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_writable(
        &self,
        region_id: RegionId,
        writable: bool
    ) -> Result<(), BoxedError>;
    fn set_readonly_gracefully<'life0, 'async_trait>(
        &'life0 self,
        region_id: RegionId
    ) -> Pin<Box<dyn Future<Output = Result<SetReadonlyResponse, BoxedError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn role(&self, region_id: RegionId) -> Option<RegionRole>;
    fn as_any(&self) -> &dyn Any;
}

Required Methods

fn name(&self) -> &str

Name of this engine

fn handle_request<'life0, 'async_trait>( &'life0 self, region_id: RegionId, request: RegionRequest ) -> Pin<Box<dyn Future<Output = Result<RegionHandleResult, BoxedError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handles non-query request to the region. Returns the count of affected rows.

fn handle_query<'life0, 'async_trait>( &'life0 self, region_id: RegionId, request: ScanRequest ) -> Pin<Box<dyn Future<Output = Result<SendableRecordBatchStream, BoxedError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handles substrait query and return a stream of record batches

fn get_metadata<'life0, 'async_trait>( &'life0 self, region_id: RegionId ) -> Pin<Box<dyn Future<Output = Result<RegionMetadataRef, BoxedError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves region’s metadata.

fn region_disk_usage<'life0, 'async_trait>( &'life0 self, region_id: RegionId ) -> Pin<Box<dyn Future<Output = Option<i64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves region’s disk usage.

fn stop<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<(), BoxedError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Stops the engine

fn set_writable( &self, region_id: RegionId, writable: bool ) -> Result<(), BoxedError>

Sets writable mode for a region.

The engine checks whether the region is writable before writing to the region. Setting the region as readonly doesn’t guarantee that write operations in progress will not take effect.

fn set_readonly_gracefully<'life0, 'async_trait>( &'life0 self, region_id: RegionId ) -> Pin<Box<dyn Future<Output = Result<SetReadonlyResponse, BoxedError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sets readonly for a region gracefully.

After the call returns, the engine ensures no more write operations will succeed in the region.

fn role(&self, region_id: RegionId) -> Option<RegionRole>

Indicates region role.

Returns the None if the region is not found.

fn as_any(&self) -> &dyn Any