Trait store_api::region_engine::RegionScanner

pub trait RegionScanner: Debug + DisplayAs + Send {
    // Required methods
    fn properties(&self) -> &ScannerProperties;
    fn schema(&self) -> SchemaRef;
    fn prepare(
        &mut self,
        ranges: Vec<Vec<PartitionRange>>
    ) -> Result<(), BoxedError>;
    fn scan_partition(
        &self,
        partition: usize
    ) -> Result<SendableRecordBatchStream, BoxedError>;
}

A scanner that provides a way to scan the region concurrently. The scanner splits the region into partitions so that each partition can be scanned concurrently. You can use this trait to implement an ExecutionPlan.

Required Methods

fn properties(&self) -> &ScannerProperties

Returns the properties of the scanner.

fn schema(&self) -> SchemaRef

Returns the schema of the record batches.

fn prepare( &mut self, ranges: Vec<Vec<PartitionRange>> ) -> Result<(), BoxedError>

Prepares the scanner with the given partition ranges.

This method is for the planner to adjust the scanner’s behavior based on the partition ranges.

fn scan_partition( &self, partition: usize ) -> Result<SendableRecordBatchStream, BoxedError>

Scans the partition and returns a stream of record batches.

Panics

Panics if the partition is out of bound.