Module store_api::storage

Storage APIs.

Modules

chunk 🔒

consts

Constants.

descriptors 🔒

engine 🔒

Storage Engine traits.

metadata 🔒

region 🔒

Region holds chunks of rows stored in the storage engine, but does not require that rows must have continuous primary key range, which is implementation specific.

requests 🔒

responses 🔒

snapshot 🔒

types 🔒

Common types.

Structs

AddColumn

Operation to add a column.

AlterRequest

Alter region request.

Chunk

A bunch of rows in columnar format.

CloseContext

CloseOptions

Options to close a region.

ColumnDescriptor

A ColumnDescriptor contains information to create a column.

ColumnDescriptorBuilder

Builder for ColumnDescriptor.

ColumnFamilyDescriptor

A ColumnFamilyDescriptor contains information to create a column family.

ColumnFamilyDescriptorBuilder

Builder for ColumnFamilyDescriptor.

ColumnSchema

Schema of a column, used as an immutable struct.

CompactContext

CreateOptions

Options to create a region.

EngineContext

Storage engine context.

FlushContext

Context for flush operations.

GetRequest

GetResponse

OpenOptions

Options to open a region.

ReadContext

Context for read.

RegionDescriptor

A RegionDescriptor contains information to create a region.

RegionDescriptorBuilder

Builder for RegionDescriptor.

RegionId

Id of the region. It’s generated by concatenating table id, region group and region number.

RegionStat

RowKeyDescriptor

A RowKeyDescriptor contains information about row key.

RowKeyDescriptorBuilder

Builder for RowKeyDescriptor.

ScanRequest

ScanResponse

Schema

A common schema, should be immutable.

SchemaBuilder

TwcsOptions

TWCS compaction options.

WriteContext

Context for write operations.

WriteResponse

Enums

AlterOperation

Operation to alter a region.

ColumnDefaultConstraint

Column’s default constraint.

ColumnDescriptorBuilderError

Error type for ColumnDescriptorBuilder

ColumnFamilyDescriptorBuilderError

Error type for ColumnFamilyDescriptorBuilder

CompactionStrategy

Options for compactions

ConcreteDataType

FlushReason

Reason of flush operation.

OpType

Operation type of the value to write to storage.

RegionDescriptorBuilderError

Error type for RegionDescriptorBuilder

RowKeyDescriptorBuilderError

Error type for RowKeyDescriptorBuilder

Constants

MAX_REGION_SEQ

The max valid region sequence number.

Traits

ChunkReader

ChunkReader is similar to async iterator of Chunk.

Region

Chunks of rows in storage engine.

RegionMeta

Metadata of a region.

Snapshot

A consistent read-only view of region.

StorageEngine

Storage engine provides primitive operations to store and access data.

WriteRequest

Write request holds a collection of updates to apply to a region.

Type Definitions

ColumnFamilyId

Id of column family. Unique in each region.

ColumnId

Id of column. Unique in each region.

RegionGroup

Group number of one region. Unique in each region.

RegionNumber

Id of regions under the same table. Unique in each table. Is composed by RegionGroup and RegionSeq.

RegionSeq

Sequence number of region inside one table. Unique in each table. The first 8 bits are preserved for RegionGroup.

SchemaRef

SequenceNumber

Represents a sequence number of data in storage. The offset of logstore can be used as a sequence number.

TableId

Id of table. Universal unique.