store_api::storage

Enum ConcreteDataType

pub enum ConcreteDataType {

    Null(NullType),
    Boolean(BooleanType),
    Int8(Int8Type),
    Int16(Int16Type),
    Int32(Int32Type),
    Int64(Int64Type),
    UInt8(UInt8Type),
    UInt16(UInt16Type),
    UInt32(UInt32Type),
    UInt64(UInt64Type),
    Float32(Float32Type),
    Float64(Float64Type),
    Decimal128(Decimal128Type),
    Binary(BinaryType),
    String(StringType),
    Date(DateType),
    Timestamp(TimestampType),
    Time(TimeType),
    Duration(DurationType),
    Interval(IntervalType),
    List(ListType),
    Dictionary(DictionaryType),
    Json(JsonType),
    Vector(VectorType),

}

Variants

Null(NullType)

Boolean(BooleanType)

Int8(Int8Type)

Int16(Int16Type)

Int32(Int32Type)

Int64(Int64Type)

UInt8(UInt8Type)

UInt16(UInt16Type)

UInt32(UInt32Type)

UInt64(UInt64Type)

Float32(Float32Type)

Float64(Float64Type)

Decimal128(Decimal128Type)

Binary(BinaryType)

String(StringType)

Date(DateType)

Timestamp(TimestampType)

Time(TimeType)

Duration(DurationType)

Interval(IntervalType)

List(ListType)

Dictionary(DictionaryType)

Json(JsonType)

Vector(VectorType)

Implementations

impl ConcreteDataType

pub fn is_float(&self) -> bool

pub fn is_boolean(&self) -> bool

pub fn is_string(&self) -> bool

pub fn is_stringifiable(&self) -> bool

pub fn is_signed(&self) -> bool

pub fn is_unsigned(&self) -> bool

pub fn is_numeric(&self) -> bool

pub fn is_timestamp(&self) -> bool

pub fn is_decimal(&self) -> bool

pub fn is_json(&self) -> bool

pub fn is_vector(&self) -> bool

pub fn numerics() -> Vec<ConcreteDataType>

pub fn unsigned_integers() -> Vec<ConcreteDataType>

pub fn timestamps() -> Vec<ConcreteDataType>

pub fn from_arrow_type(dt: &DataType) -> ConcreteDataType

Convert arrow data type to ConcreteDataType.

Panics

Panic if given arrow data type is not supported.

pub fn is_null(&self) -> bool

pub fn as_list(&self) -> Option<&ListType>

Try to cast the type as a [ListType].

pub fn as_timestamp(&self) -> Option<TimestampType>

Try to cast data type as a [TimestampType].

pub fn numeric_precision(&self) -> Option<u8>

Try to get numeric precision, returns None if it’s not numeric type

pub fn numeric_scale(&self) -> Option<i8>

Try to get numeric scale, returns None if it’s float or not numeric type

pub fn as_time(&self) -> Option<TimeType>

Try to cast data type as a [TimeType].

pub fn as_decimal128(&self) -> Option<Decimal128Type>

pub fn as_json(&self) -> Option<JsonType>

pub fn as_vector(&self) -> Option<VectorType>

pub fn can_arrow_type_cast_to(&self, to_type: &ConcreteDataType) -> bool

Checks if the data type can cast to another data type.

pub fn as_duration(&self) -> Option<DurationType>

Try to cast data type as a [DurationType].

pub fn postgres_datatype_name(&self) -> &'static str

Return the datatype name in postgres type system

impl ConcreteDataType

pub fn null_datatype() -> ConcreteDataType

pub fn boolean_datatype() -> ConcreteDataType

pub fn uint8_datatype() -> ConcreteDataType

pub fn uint16_datatype() -> ConcreteDataType

pub fn uint32_datatype() -> ConcreteDataType

pub fn uint64_datatype() -> ConcreteDataType

pub fn int8_datatype() -> ConcreteDataType

pub fn int16_datatype() -> ConcreteDataType

pub fn int32_datatype() -> ConcreteDataType

pub fn int64_datatype() -> ConcreteDataType

pub fn float32_datatype() -> ConcreteDataType

pub fn float64_datatype() -> ConcreteDataType

pub fn binary_datatype() -> ConcreteDataType

pub fn date_datatype() -> ConcreteDataType

pub fn string_datatype() -> ConcreteDataType

pub fn json_datatype() -> ConcreteDataType

impl ConcreteDataType

pub fn timestamp_second_datatype() -> ConcreteDataType

pub fn timestamp_millisecond_datatype() -> ConcreteDataType

pub fn timestamp_microsecond_datatype() -> ConcreteDataType

pub fn timestamp_nanosecond_datatype() -> ConcreteDataType

pub fn time_datatype(unit: TimeUnit) -> ConcreteDataType

Returns the time data type with TimeUnit.

pub fn time_second_datatype() -> ConcreteDataType

Creates a [Time(TimeSecondType)] datatype.

pub fn time_millisecond_datatype() -> ConcreteDataType

Creates a [Time(TimeMillisecondType)] datatype.

pub fn time_microsecond_datatype() -> ConcreteDataType

Creates a [Time(TimeMicrosecond)] datatype.

pub fn time_nanosecond_datatype() -> ConcreteDataType

Creates a [Time(TimeNanosecond)] datatype.

pub fn duration_second_datatype() -> ConcreteDataType

Creates a [Duration(DurationSecondType)] datatype.

pub fn duration_millisecond_datatype() -> ConcreteDataType

Creates a [Duration(DurationMillisecondType)] datatype.

pub fn duration_microsecond_datatype() -> ConcreteDataType

Creates a [Duration(DurationMicrosecondType)] datatype.

pub fn duration_nanosecond_datatype() -> ConcreteDataType

Creates a [Duration(DurationNanosecondType)] datatype.

pub fn interval_month_day_nano_datatype() -> ConcreteDataType

Creates a [Interval(IntervalMonthDayNanoType)] datatype.

pub fn interval_year_month_datatype() -> ConcreteDataType

Creates a [Interval(IntervalYearMonthType)] datatype.

pub fn interval_day_time_datatype() -> ConcreteDataType

Creates a [Interval(IntervalDayTimeType)] datatype.

pub fn timestamp_datatype(unit: TimeUnit) -> ConcreteDataType

pub fn from_arrow_time_unit(t: &TimeUnit) -> ConcreteDataType

Converts from arrow timestamp unit to

pub fn duration_datatype(unit: TimeUnit) -> ConcreteDataType

pub fn interval_datatype(unit: IntervalUnit) -> ConcreteDataType

pub fn from_arrow_interval_unit(u: &IntervalUnit) -> ConcreteDataType

pub fn list_datatype(item_type: ConcreteDataType) -> ConcreteDataType

pub fn dictionary_datatype( key_type: ConcreteDataType, value_type: ConcreteDataType, ) -> ConcreteDataType

pub fn decimal128_datatype(precision: u8, scale: i8) -> ConcreteDataType

pub fn decimal128_default_datatype() -> ConcreteDataType

pub fn vector_datatype(dim: u32) -> ConcreteDataType

pub fn vector_default_datatype() -> ConcreteDataType