Trait store_api::storage::StorageEngine

pub trait StorageEngine: Send + Sync + Clone + 'static {
    type Error: ErrorExt + Send + Sync;
    type Region: Region;

    // Required methods
    fn open_region<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        ctx: &'life1 EngineContext,
        name: &'life2 str,
        opts: &'life3 OpenOptions
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Region>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn close_region<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ctx: &'life1 EngineContext,
        region: Self::Region
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn create_region<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        ctx: &'life1 EngineContext,
        descriptor: RegionDescriptor,
        opts: &'life2 CreateOptions
    ) -> Pin<Box<dyn Future<Output = Result<Self::Region, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn drop_region<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ctx: &'life1 EngineContext,
        region: Self::Region
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_region(
        &self,
        ctx: &EngineContext,
        name: &str
    ) -> Result<Option<Self::Region>, Self::Error>;
}

Storage engine provides primitive operations to store and access data.

Required Associated Types

type Error: ErrorExt + Send + Sync

type Region: Region

Required Methods

fn open_region<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, ctx: &'life1 EngineContext, name: &'life2 str, opts: &'life3 OpenOptions ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Region>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Opens an existing region. Returns Ok(None) if region does not exists.

fn close_region<'life0, 'life1, 'async_trait>( &'life0 self, ctx: &'life1 EngineContext, region: Self::Region ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Closes given region.

fn create_region<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, ctx: &'life1 EngineContext, descriptor: RegionDescriptor, opts: &'life2 CreateOptions ) -> Pin<Box<dyn Future<Output = Result<Self::Region, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Creates and returns the created region.

Returns existing region if region with same name already exists. The region will be opened before returning.

fn drop_region<'life0, 'life1, 'async_trait>( &'life0 self, ctx: &'life1 EngineContext, region: Self::Region ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Drops given region.

The region will be closed before dropping.

fn get_region( &self, ctx: &EngineContext, name: &str ) -> Result<Option<Self::Region>, Self::Error>

Returns the opened region with given name.