Crate store_api

Storage related APIs

Modules

data_source

logstore

LogStore APIs.

manifest

metadata service

metadata

Metadata of region and column.

metric_engine_consts

Constants used in metric engine.

mito_engine_options

Option keys for the mito engine. We define them in this mod so the create parser can use it to validate table options.

path_utils

Path constants for table engines, cluster states and WAL All paths relative to data_home(file storage) or root path(S3, OSS etc).

region_engine

Region Engine’s definition

region_request

storage

Storage APIs.