Trait store_api::logstore::LogStore

pub trait LogStore: Send + Sync + 'static + Debug {
    type Error: ErrorExt + Send + Sync + 'static;
    type Namespace: Namespace;
    type Entry: Entry;

    // Required methods
    fn stop<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn append<'life0, 'async_trait>(
        &'life0 self,
        e: Self::Entry
    ) -> Pin<Box<dyn Future<Output = Result<AppendResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn append_batch<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ns: &'life1 Self::Namespace,
        e: Vec<Self::Entry>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Id>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn read<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ns: &'life1 Self::Namespace,
        id: Id
    ) -> Pin<Box<dyn Future<Output = Result<SendableEntryStream<'_, Self::Entry, Self::Error>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn create_namespace<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ns: &'life1 Self::Namespace
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn delete_namespace<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ns: &'life1 Self::Namespace
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn list_namespaces<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Namespace>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn entry<D: AsRef<[u8]>>(
        &self,
        data: D,
        id: Id,
        ns: Self::Namespace
    ) -> Self::Entry;
    fn namespace(&self, id: Id) -> Self::Namespace;
    fn obsolete<'life0, 'async_trait>(
        &'life0 self,
        namespace: Self::Namespace,
        id: Id
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

LogStore serves as a Write-Ahead-Log for storage engine.

Required Associated Types

type Error: ErrorExt + Send + Sync + 'static

type Namespace: Namespace

type Entry: Entry

Required Methods

fn stop<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Stop components of logstore.

fn append<'life0, 'async_trait>( &'life0 self, e: Self::Entry ) -> Pin<Box<dyn Future<Output = Result<AppendResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Append an Entry to WAL with given namespace and return append response containing the entry id.

fn append_batch<'life0, 'life1, 'async_trait>( &'life0 self, ns: &'life1 Self::Namespace, e: Vec<Self::Entry> ) -> Pin<Box<dyn Future<Output = Result<Vec<Id>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Append a batch of entries atomically and return the offset of first entry.

fn read<'life0, 'life1, 'async_trait>( &'life0 self, ns: &'life1 Self::Namespace, id: Id ) -> Pin<Box<dyn Future<Output = Result<SendableEntryStream<'_, Self::Entry, Self::Error>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create a new EntryStream to asynchronously generates Entry with ids starting from id.

fn create_namespace<'life0, 'life1, 'async_trait>( &'life0 mut self, ns: &'life1 Self::Namespace ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create a new Namespace.

fn delete_namespace<'life0, 'life1, 'async_trait>( &'life0 mut self, ns: &'life1 Self::Namespace ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Delete an existing Namespace with given ref.

fn list_namespaces<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Namespace>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

List all existing namespaces.

fn entry<D: AsRef<[u8]>>( &self, data: D, id: Id, ns: Self::Namespace ) -> Self::Entry

Create an entry of the associate Entry type

fn namespace(&self, id: Id) -> Self::Namespace

Create a namespace of the associate Namespace type

fn obsolete<'life0, 'async_trait>( &'life0 self, namespace: Self::Namespace, id: Id ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Mark all entry ids <=id of given namespace as obsolete so that logstore can safely delete the log files if all entries inside are obsolete. This method may not delete log files immediately.