Module store_api::manifest::action

Common actions for manifest

Structs

ProtocolAction

Protocol action that used to block older clients from reading or writing the log when backwards incompatible changes are made to the protocol. clients should be tolerant of messages and fields that they do not understand.

VersionHeader

Constants

READER_VERSION 🔒

Current reader and writer versions TODO(dennis): configurable

WRITER_VERSION 🔒

Functions

supported_protocol_version

The maximum protocol version we are currently allowed to use, TODO(dennis): reading from configuration.

Type Aliases

ProtocolVersion