Module store_api::metric_engine_consts

Constants used in metric engine.

Constants

ALTER_PHYSICAL_EXTENSION_KEY

HashMap key to be used in the region server’s extension response. Represent a list of column metadata that are added to physical table.

DATA_REGION_SUBDIR

DATA_SCHEMA_TABLE_ID_COLUMN_NAME

Column name of internal column __metric that stores the original metric name

DATA_SCHEMA_TSID_COLUMN_NAME

FILE_ENGINE_NAME

LOGICAL_TABLE_METADATA_KEY

Metadata key present in the CREATE TABLE ... WITH () clause. This key is used to identify a logical table and associate it with a corresponding physical table . E.g.:

METADATA_REGION_SUBDIR

METADATA_SCHEMA_KEY_COLUMN_INDEX

METADATA_SCHEMA_KEY_COLUMN_NAME

METADATA_SCHEMA_TIMESTAMP_COLUMN_INDEX

METADATA_SCHEMA_TIMESTAMP_COLUMN_NAME

METADATA_SCHEMA_VALUE_COLUMN_INDEX

METADATA_SCHEMA_VALUE_COLUMN_NAME

METRIC_DATA_REGION_GROUP

region group value for data region inside a metric region

METRIC_ENGINE_NAME

METRIC_METADATA_REGION_GROUP

region group value for metadata region inside a metric region

PHYSICAL_TABLE_METADATA_KEY

Metadata key present in the CREATE TABLE ... WITH () clause. This key is used to identify the table is a physical metric table. E.g.:

Functions

is_metric_engine

Returns true if it’s metric engine

is_metric_engine_internal_column

Returns true if it’s a internal column of the metric engine.