store_api

Module mito_engine_options

Option keys for the mito engine. We define them in this mod so the create parser can use it to validate table options.

Constants

APPEND_MODE_KEY

Option key for append mode.

COMPACTION_TYPE

Option key for compaction type.

COMPACTION_TYPE_TWCS

TWCS compaction strategy.

MEMTABLE_PARTITION_TREE_DATA_FREEZE_THRESHOLD

Option key for memtable partition tree data freeze threshold.

MEMTABLE_PARTITION_TREE_FORK_DICTIONARY_BYTES

Option key for memtable partition tree fork dictionary bytes.

MEMTABLE_PARTITION_TREE_INDEX_MAX_KEYS_PER_SHARD

Option key for memtable partition tree index max keys per shard.

MEMTABLE_PARTITION_TREE_PRIMARY_KEY_ENCODING

Option key for memtable partition tree primary key encoding.

MEMTABLE_TYPE

Option key for memtable type.

MERGE_MODE_KEY

Option key for merge mode.

REMOTE_COMPACTION

Option key for twcs remote compaction.

SKIP_WAL_KEY

Option key for skipping WAL.

SNAPSHOT_READ

Option key for snapshot read.

TTL_KEY

Option key for TTL(time-to-live)

TWCS_FALLBACK_TO_LOCAL

Option key for twcs fallback to local.

TWCS_MAX_OUTPUT_FILE_SIZE

Option key for twcs max output file size.

TWCS_TIME_WINDOW

Option key for twcs time window.

TWCS_TRIGGER_FILE_NUM

Option key for twcs min file num to trigger a compaction.

Functions

is_mito_engine_option_key

Returns true if the key is a valid option key for the mito engine.