store_api

Module storage

Storage APIs.

Modules

consts

Constants.

descriptors 🔒

requests 🔒

types 🔒

Common types.

Structs

ColumnDescriptor

A ColumnDescriptor contains information to create a column.

ColumnDescriptorBuilder

Builder for ColumnDescriptor.

ColumnSchema

Schema of a column, used as an immutable struct.

RegionId

Id of the region. It’s generated by concatenating table id, region group and region number.

ScanRequest

Schema

A common schema, should be immutable.

SchemaBuilder

Enums

ColumnDefaultConstraint

Column’s default constraint.

ColumnDescriptorBuilderError

Error type for ColumnDescriptorBuilder

ConcreteDataType

TimeSeriesDistribution

A hint on how to distribute time-series data on the scan output.

TimeSeriesRowSelector

A hint on how to select rows from a time-series.

Constants

MAX_REGION_SEQ

The max valid region sequence number.

Type Aliases

ColumnId

Id of column. Unique in each region.

RegionGroup

Group number of one region. Unique in each region.

RegionNumber

Id of regions under the same table. Unique in each table. Is composed by RegionGroup and RegionSeq.

RegionSeq

Sequence number of region inside one table. Unique in each table. The first 8 bits are preserved for RegionGroup.

SchemaRef

SequenceNumber

Represents a sequence number of data in storage. The offset of logstore can be used as a sequence number.

TableId

Id of table. Universal unique.