Module store_api::storage::engine

Storage Engine traits.

StorageEngine is the abstraction over a multi-regions, schematized data storage system, a StorageEngine instance manages a bunch of storage unit called Region, which holds chunks of rows, support operations like PUT/DELETE/SCAN.

Structs

CloseOptions

Options to close a region.

CreateOptions

Options to create a region.

EngineContext

Storage engine context.

OpenOptions

Options to open a region.

TwcsOptions

TWCS compaction options.

Enums

CompactionStrategy

Options for compactions

Constants

COMPACTION_STRATEGY_KEY 🔒

COMPACTION_STRATEGY_TWCS_VALUE 🔒

TWCS_MAX_ACTIVE_WINDOW_FILES_KEY 🔒

TWCS_MAX_INACTIVE_WINDOW_FILES_KEY 🔒

TWCS_TIME_WINDOW_SECONDS_KEY 🔒

Traits

StorageEngine

Storage engine provides primitive operations to store and access data.